+++ app/simple_pages/__init__.py
@@ -1,0 +1,1 @@
+from . import routes
