+++ app/templates/cookies/cookies.html
