+++ app/templates/registration.html
